auth guard

+++ backend/src/resolvers/kittens.ts
@@ -1,10 +1,8 @@
-/* eslint-disable @typescript-eslint/no-unsafe-member-access */
-/* eslint-disable @typescript-eslint/no-unsafe-assignment */
 import Kitten from "../models/kittens.js";
+import { checkAuth } from "../utils/checkAuth.ts";
 
 const kittensQuery = {
   getKitten: async (_, args) => {
-    // console.log("From getKittens!");
     return await Kitten.findOne({ _id: args.id });
   },
   getKittens: async () => {
@@ -13,20 +11,49 @@ const kittensQuery = {
 };
 
 const kittensMutation = {
-  createKitten: async (_, args) => {
-    const kitten = new Kitten({
-      name: args.name,
-      breed: args.breed,
-    });
+  createKitten: async (parent, args, req, info) => {
+    try {
+      const auth = await checkAuth(req.headers.authorization);
+      const kitten = new Kitten({
+        name: args.name,
+        breed: args.breed,
+        userId: auth._id.toString(),
+      });
 
-    return await kitten.save();
+      return await kitten.save();
+    } catch (err) {
+      console.error(err);
+      throw new Error(err.message);
+    }
   },
-  updateKitten: async (parent, args) => {
-    await Kitten.findOneAndUpdate({ _id: args.id }, args as object);
-    return await Kitten.findOne({ _id: args.id });
+  updateKitten: async (_, args, req) => {
+    try {
+      const auth = await checkAuth(req.headers.authorization);
+
+      if (auth.role === "admin") {
+        await Kitten.findOneAndUpdate({ _id: args.id }, args as object);
+        return await Kitten.findOne({ _id: args.id });
+      } else {
+        throw new Error("Not authorized!");
+      }
+    } catch (err) {
+      console.error(err);
+      throw new Error(err.message);
+    }
   },
-  deleteKitten: async (parent, args) => {
-    return await Kitten.findOneAndDelete({ _id: args.id });
+  deleteKitten: async (_, args, req) => {
+    try {
+      const auth = await checkAuth(req.headers.authorization);
+
+      if (auth.role === "admin") {
+        return await Kitten.findOneAndDelete({ _id: args.id });
+      } else {
+        throw new Error("Not authorized!");
+      }
+    } catch (err) {
+      console.error(err);
+      throw new Error(err.message);
+    }
   },
 };
